fix(blog-front) :star: including headers in request
The request will include headers that pertain the user id. This header will then be used by the backend to create a relation between the blog and the user.

As of now, this is working as a custom header. However, this should be an Authorization that is done with the user.id and the token. However, this functionality is not working at the moment, and will be dealt with at another commit